+++ tests/actions/OrderAction/updateActionTest.php
@@ -51,7 +51,7 @@ class updateActionTest extends BaseOrderAction
         $resource_id = 'pay_azerty123456';
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [],
+                'getBy' => [],
             ]);
         $this->assertSame(
             [
@@ -67,7 +67,7 @@ class updateActionTest extends BaseOrderAction
         $resource_id = 'pay_azerty123456';
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -114,7 +114,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -160,7 +160,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -211,7 +211,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -267,7 +267,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -332,7 +332,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -397,7 +397,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -472,7 +472,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',
@@ -549,7 +549,7 @@ class updateActionTest extends BaseOrderAction
             ]);
         $this->payment_repository
             ->shouldReceive([
-                'getByResourceId' => [
+                'getBy' => [
                     'id_payplug_payment' => 42,
                     'resource_id' => 'pay_azerty1234',
                     'method' => 'standard',