+++ changelog.md
@@ -1,5 +1,13 @@
 # Payplug module changelog
 
+## Version 4.11.0
+- Refactoring :
+  - [PRE-2511](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2491): Set EntityRepository
+  - [PRE-2515](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2492): Set Logger entity
+  - [PRE-2517](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2494): Set State Entity
+  - [PRE-2512](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2493): Set CacheEntity
+  - [PRE-2518](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2502): Set Payment Entity
+
 ## Version 4.10.0
 - Improvment :
   - [PRE-2551](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2496): Add order reference to payment resource